# Improved performance of utf8 comparison (1.7x-4x)