--- a/benches/filter_kernels.rs
+++ b/benches/filter_kernels.rs
@@ -113,7 +113,7 @@ fn add_benchmark(c: &mut Criterion) {
         b.iter(|| bench_built_filter(&sparse_filter, &data_array))
     });
 
-    let data_array = create_string_array::<i32>(size, 0.5);
+    let data_array = create_string_array::<i32>(size, 0.5, 42);
     c.bench_function("filter context string", |b| {
         b.iter(|| bench_built_filter(&filter, &data_array))
     });