--- a/src/util/bench_util.rs
+++ b/src/util/bench_util.rs
@@ -95,15 +95,15 @@ where
 }
 
 /// Creates an random (but fixed-seeded) array of a given size and null density
-pub fn create_string_array<O: Offset>(size: usize, null_density: f32) -> Utf8Array<O> {
-    let rng = &mut seedable_rng();
+pub fn create_string_array<O: Offset>(size: usize, null_density: f32, seed: u64) -> Utf8Array<O> {
+    let mut rng = StdRng::seed_from_u64(seed);
 
     (0..size)
         .map(|_| {
             if rng.gen::<f32>() < null_density {
                 None
             } else {
-                let value = rng
+                let value = (&mut rng)
                     .sample_iter(&Alphanumeric)
                     .take(4)
                     .map(char::from)