```
utf8 2^20               time:   [10.376 ms 10.395 ms 10.424 ms]                      
                        change: [-43.339% -43.191% -43.033%] (p = 0.00 < 0.05)
utf8 scalar 2^20        time:   [1.5213 ms 1.5243 ms 1.5283 ms]                              
                        change: [-75.416% -75.313% -75.200%] (p = 0.00 < 0.05)
```

The idea is that we can iterate directly over values (`&str`) ignoring validities, which is evidently faster than iterating over `Option<&str>`